Migrate to .NET 10, refactor online matching to in-memory, rewrite relay server
Major changes across multiple areas:

.NET 10 Migration & Build Fixes:
- Upgrade all projects to .NET 10 / C# 14
- Fix Mapperly type mismatches (int -> long in DTOs)
- Fix MudBlazor 9.x breaking changes (generic components, IMudDialogInstance,
  Href rename, nullable DialogResult, MudPopoverProvider)
- Remove obsolete Quartz UseMicrosoftDependencyInjectionJobFactory
- Enable static web assets for exe-direct execution

Library Replacements:
- Replace ChoETL with System.Xml.Serialization (XmlSerializationExtensions rewrite)
- Replace Swan.Logging with Serilog in GCRelayServer
- Add XML serialization test project (14 tests)

Architecture Refinements:
- Add GameProtocolResult custom IActionResult for game protocol envelope
- Add XmlSerializer cache (ConcurrentDictionary) to prevent memory leaks
- Set NoTracking on read-only MusicDbContext
- Fix O(n^2) lookups in rank jobs with Dictionary-based lookups
- Fix SQL-side aggregation in UpdatePlayNumRankJob (GroupBy instead of cross-join)
- Propagate Quartz cancellation tokens in all jobs
- Fix SetPlayOptionCommand count check (count == 1 -> count > 0)
- Compiled regex in ChangePlayerNameDialog
- Delete TestController (bypassed IMusicDbContext interface)

Online Matching Refactor (in-memory state):
- Replace SQLite-backed matching with singleton OnlineMatchingService
- ConcurrentDictionary<long, MatchRoom> with per-room Lock (C# 13+)
- Atomic TryAddEntry eliminates TOCTOU race on room capacity
- Immutable MatchRoomSnapshot for lock-free XML serialization
- Remove OnlineMatch/OnlineMatchEntry DB entities and tables (migration)
- New CleanupOnlineMatchesJob (Quartz, 5-min interval)
- Fix 12-hour clock bug (hh -> HH) in TimeHelper and OnlineMatchMapper

GCRelayServer Rewrite:
- Thread-safe Room class with Lock per room
- RoomManager with ConcurrentDictionary.AddOrUpdate for atomic room creation
- Try/catch on packet deserialization (no more crashes on malformed packets)
- ReceiveAsync in finally block AND OnError (server survives socket errors)
- Fix SendPacket shared packet mutation (save/restore RequestSubType)
- Graceful Ctrl+C shutdown, port validation, Log.CloseAndFlush in finally
- Delete DictEntry.cs, replace with Room.cs + RoomManager.cs

+++ Application/Dto/Game/CardBDatumDto.cs
@@ -9,5 +9,5 @@ public class CardBDatumDto
     public string CardBdata { get; set; } = string.Empty;
     
     [XmlElement(ElementName = "bdata_size")]
-    public int BDataSize { get; set; }
+    public long BDataSize { get; set; }
 }
